Boat pathing: Properly handle "goto" for water units (#456)
I noticed that while I could move a boat around with the arrow keys, I couldn't
move it with "G" or the "goto" button. Since my laptop doesn't have a numpad,
this made it tricky to explore.

I think this is a fairly straightforward fix to the problem:

1. I added an implementation of the EdgeWalker interface that handles water tiles,
   including cities in the set of neighbors to allow canal cities, and pathing
   back into coastal cities.

2. Require specifying land vs water units when picking a pathing algorithm.

I did some manual testing, including building a canal city north of the default
spawn, and confirmed I could path through it. I think building the city on a
hill amusingly resulted in the boat movement cost being 2, but besides that
it worked.