Introduce a new layer for displaying the "goto" cursor
This is a nice "quality of life" feature that also helped me identify
bugs in the movement cost calculations, such as the one fixed in [this
PR](#457).

This doesn't yet implement a line or some other mechanism to show the
actual planned path, but it both displays the tile that the user wants
to go to, and the cost of getting there in turns.

I moved the code for calculating the cost between two tiles into
`TilePath.cs`, since that's a lower dependency file, and this sort of
common utility should probably be in a low dependency file.