Fix unresponsive unit buttons
Prior to this commit, some of the unit buttons (e.g., Build City,
Build Road) could become unresponsive. This issue was related to how
the game handled button presses. When a button was pressed, the game
emulated the press of an action hotkey, which should have been handled
by the Game class. However, in some cases, the input simulation didn't work
properly, and the game didn't react to the button presses.

To solve this issue, I decoupled the handling of button presses from
the handling of hotkey input. Now, instead of imitating input, a unit
button sends a signal to the Game class. In the Game class, the logic for
handling action requests has been moved into a new method, which is
now used for both handling signals from unit buttons and processing
hotkey input.