Display mines in loaded save files and scenarios
This doesn't yet add the worker action to create new mines (I wanted to
have a single PR that just show how to add a tile overlay, and a second
one just to show the worker action).

Rather than add a new file in `Map/` for mine overlays, I renamed
`RoadLayer` to `TileOverlayLayer`, so that all the various tile overlay
graphics logic can be in one place. This also makes it easy to ensure
that the mines are drawn after the roads/railroads.