UnitInteractions: Return a list of unit actions, not a unit object
This accomplishes the TODO mentioned in the function comment and will
simplify some upcoming work to add a mine button.

The mine button is different from existing buttons in that it changes
based on the terrain *and* that can happen within a single turn. While
the existing road button also depends on the terrain and whether it has
a road, any time we move a worker to a tile without a road it ends the
worker's turn, so this doesn't matter.

An upcoming PR will need to send the appropriate signal to update the
buttons after a unit stops walking in case the unit still has moves.